+++ CHANGELOG.md
@@ -22,6 +22,7 @@ CHANGELOG
  * Allow Usages to be specified via `#[AsCommand]` attribute.
  * Allow passing invokable commands to `Symfony\Component\Console\Tester\CommandTester`
  * Add `#[Input]` attribute to support DTOs in commands
+ * Add optional timeout for interaction in `QuestionHelper`
 
 7.3
 ---